+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/ArrayValueEditor.java
@@ -168,7 +168,7 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 	}
 
 	@Override
-	public void setToOverride(boolean override) {
+	public void setToCustomData(boolean override) {
 		masterPane.getChildren().clear();
 		if (override) {
 			masterPane.getChildren().add(overrideField);