+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/BooleanValueEditor.java
@@ -60,7 +60,7 @@ public class BooleanValueEditor implements ValueEditor<SVBoolean> {
 	}
 
 	@Override
-	public void setToOverride(boolean override) {
+	public void setToCustomData(boolean override) {
 		masterPane.getChildren().clear();
 		if (override) {
 			masterPane.getChildren().add(overrideField);