+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/ColorValueEditor.java
@@ -74,7 +74,7 @@ public class ColorValueEditor implements ValueEditor<AColor> {
 	}
 	
 	@Override
-	public void setToOverride(boolean override) {
+	public void setToCustomData(boolean override) {
 		masterPane.getChildren().clear();
 		if (override) {
 			masterPane.getChildren().add(overrideField);