+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/FontValueEditor.java
@@ -75,7 +75,7 @@ public class FontValueEditor implements ValueEditor<AFont> {
 	}
 
 	@Override
-	public void setToOverride(boolean override) {
+	public void setToCustomData(boolean override) {
 		masterPane.getChildren().clear();
 		if(override){
 			masterPane.getChildren().add(overrideField);