+++ src/com/kaylerrenslow/armaDialogCreator/util/UpdateListenerGroup.java
@@ -10,32 +10,30 @@
 
 package com.kaylerrenslow.armaDialogCreator.util;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.LinkedList;
-import java.util.List;
 
 /**
  Created by Kayler on 07/05/2016.
  */
 public class UpdateListenerGroup<T> {
-	private final ObservableList<UpdateGroupListener<T>> updateListeners = FXCollections.observableArrayList(new LinkedList<>());
-	private final List<UpdateListenerGroup<T>> chain = new LinkedList<>();
-
-	@NotNull
-	public ObservableList<UpdateGroupListener<T>> getUpdateListeners() {
-		return updateListeners;
-	}
+	private final LinkedList<UpdateGroupListener<T>> updateListeners = new LinkedList<>();
+	private final LinkedList<UpdateListenerGroup<T>> chain = new LinkedList<>();
+	private final LinkedList<UpdateGroupListener<T>> newListeners = new LinkedList<>();
+	private boolean iterating = false; //prevent CoModificationException
 
 	/** Will add the given listener. If the listener has already been added, will do nothing (no duplicates allowed). */
 	public void addListener(@NotNull UpdateGroupListener<T> listener) {
 		if (updateListeners.contains(listener)) {
 			return;
 		}
-		updateListeners.add(listener);
+		if (iterating) {
+			newListeners.add(listener);
+		} else {
+			updateListeners.add(listener);
+		}
 	}
 
 	public boolean removeListener(@NotNull UpdateGroupListener<T> listener) {
@@ -43,11 +41,16 @@ public class UpdateListenerGroup<T> {
 	}
 
 	public void update(@Nullable T data) {
+		iterating = true;
 		for (UpdateGroupListener<T> updateListener : updateListeners) {
 			updateListener.update(this, data);
 		}
 		for (UpdateListenerGroup<T> group : chain) {
 			group.update(data);
+		}
+		iterating = false;
+		while (newListeners.size() > 0) {
+			updateListeners.add(newListeners.pop());
 		}
 	}